refactor(bridge): extract runtime handoff loop
