+++ ARCHITECTURE.md
@@ -706,6 +706,7 @@ src/
 │   │   │   ├── recovery_text.rs
 │   │   │   ├── response_delivery.rs
 │   │   │   ├── retry_state.rs
+│   │   │   ├── runtime_handoff_loop.rs
 │   │   │   ├── single_message_footer.rs
 │   │   │   ├── skill_usage.rs
 │   │   │   ├── stale_resume.rs
