+++ src/services/discord/turn_bridge/mod.rs
@@ -15,6 +15,7 @@ mod panel_lifecycle;
 mod recall_feedback;
 pub(in crate::services::discord) mod recovery_text;
 mod retry_state;
+mod runtime_handoff_loop;
 mod single_message_footer;
 mod skill_usage;
 mod stale_resume;
@@ -61,9 +62,7 @@ use crate::db::session_observability::{
 };
 use crate::db::session_transcripts::{SessionTranscriptEvent, SessionTranscriptEventKind};
 use crate::db::turns::TurnTokenUsage;
-use crate::services::agent_protocol::{
-    RuntimeHandoff, RuntimeHandoffKind, StatusEvent, TaskNotificationKind,
-};
+use crate::services::agent_protocol::{StatusEvent, TaskNotificationKind};
 use crate::services::memory::{
     CaptureRequest, TokenUsage, resolve_memory_role_id, resolve_memory_session_id,
 };
@@ -197,6 +196,10 @@ use retry_state::{
     clear_local_session_state, handle_gemini_retry_boundary, reset_session_for_auto_retry,
     rewind_and_persist_delivery_on_reclaim, sync_response_delivery_state,
     sync_terminal_error_delivery_state_for_bridge_owner,
+};
+use runtime_handoff_loop::{
+    RuntimeHandoffLoopContext, RuntimeHandoffLoopMessage, RuntimeHandoffLoopOutcome,
+    RuntimeHandoffLoopState, handle_runtime_handoff_loop_message,
 };
 use skill_usage::record_skill_usage_from_tool_use;
 use stale_resume::{
@@ -934,309 +937,6 @@ fn maybe_refresh_active_turn_activity_heartbeat_at(
     _last_heartbeat_at: &mut Option<std::time::Instant>,
     _now: std::time::Instant,
 ) {
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_watcher_runtime_handoff(
-    shared_owned: &Arc<SharedData>,
-    provider: &ProviderKind,
-    channel_id: ChannelId,
-    inflight_state: &mut InflightTurnState,
-    runtime_kind: RuntimeHandoffKind,
-    output_path: String,
-    input_fifo_path: Option<String>,
-    tmux_session_name: String,
-    last_offset: u64,
-    tmux_last_offset: &mut Option<u64>,
-    watcher_owner_channel_id: &mut ChannelId,
-    standby_relay_owns_output: &mut bool,
-    watcher_relay_available_for_turn: &mut bool,
-    watcher_handoff_claim_outcome: &mut WatcherHandoffClaimOutcome,
-    tmux_handed_off: &mut bool,
-    watcher_owns_assistant_relay: &mut bool,
-    state_dirty: &mut bool,
-    done: bool,
-    terminal_control_drain_until: &mut Option<std::time::Instant>,
-) {
-    *tmux_last_offset = Some(last_offset);
-    inflight_state.runtime_kind = Some(runtime_kind);
-    inflight_state.tmux_session_name = Some(tmux_session_name.clone());
-    inflight_state.output_path = Some(output_path.clone());
-    let mut fifo_path = input_fifo_path.filter(|path| !path.is_empty());
-    // #2235 one-release compat window: ClaudeTui rows must still ship a
-    // populated `input_fifo_path` so a rollback to an old binary can satisfy
-    // its FIFO-required recovery branch. Synthesize from the canonical
-    // per-session tmux path when the caller didn't supply one.
-    if matches!(runtime_kind, RuntimeHandoffKind::ClaudeTui) && fifo_path.is_none() {
-        let (_, synthesized_fifo) = tmux_runtime_paths(&tmux_session_name);
-        if !synthesized_fifo.is_empty() {
-            fifo_path = Some(synthesized_fifo);
-        }
-    }
-    inflight_state.input_fifo_path = fifo_path;
-    inflight_state.last_offset = last_offset;
-    *state_dirty |= inflight_state.set_watcher_owner_channel_id(watcher_owner_channel_id.get());
-    // #2235 NOTE: we deliberately do NOT durably save the row here.
-    // `watcher_owns_live_relay` is still `false` at this point and only flips
-    // to `true` after the watcher is successfully claimed and spawned (the
-    // leader-branch path below). A save before that flag is set would leak a
-    // v8 row with the new handoff shape alongside `watcher_owns_live_relay =
-    // false`, which on restart would make the restored watcher yield to a
-    // phantom bridge owner (codex adversarial review on #2235). The
-    // existing branch-specific saves at the post-flag flip points plus the
-    // centralized `state_dirty` flush already cover the durable-stamp
-    // guarantee for watcher-owned RuntimeReady paths.
-    //
-    // #2263: the standby branch is INTENTIONALLY not covered by this
-    // invariant — see the in-branch comment near the
-    // `*standby_relay_owns_output = true` assignment for why the flag
-    // stays `false` on standby and the trade-off vs duplicate delivery.
-
-    // #226: Atomic claim via try_claim_watcher
-    let cancel = Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let paused = Arc::new(std::sync::atomic::AtomicBool::new(true));
-    let resume_offset = Arc::new(std::sync::Mutex::new(None::<u64>));
-    let pause_epoch = Arc::new(std::sync::atomic::AtomicU64::new(1));
-    let turn_delivered = Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let last_heartbeat_ts_ms = Arc::new(std::sync::atomic::AtomicI64::new(
-        super::tmux_watcher_now_ms(),
-    ));
-    let handle = TmuxWatcherHandle {
-        tmux_session_name: tmux_session_name.clone(),
-        output_path: output_path.clone(),
-        paused: paused.clone(),
-        resume_offset: resume_offset.clone(),
-        cancel: cancel.clone(),
-        pause_epoch: pause_epoch.clone(),
-        turn_delivered: turn_delivered.clone(),
-        last_heartbeat_ts_ms: last_heartbeat_ts_ms.clone(),
-    };
-    #[cfg(unix)]
-    let (watcher_claimed, watcher_claim_replaced_existing) = {
-        let claim = super::tmux::claim_or_reuse_watcher(
-            &shared_owned.tmux_watchers,
-            channel_id,
-            handle,
-            provider,
-            "turn_bridge_runtime_ready",
-        );
-        *watcher_owner_channel_id = claim.owner_channel_id();
-        *state_dirty |= inflight_state.set_watcher_owner_channel_id(watcher_owner_channel_id.get());
-        (claim.should_spawn(), claim.replaced_existing())
-    };
-    #[cfg(not(unix))]
-    let (watcher_claimed, watcher_claim_replaced_existing) = {
-        let _ = handle;
-        (false, false)
-    };
-    #[cfg(unix)]
-    let mut watcher_ready_for_relay = !watcher_claimed;
-    #[cfg(not(unix))]
-    let mut watcher_ready_for_relay = false;
-    *watcher_handoff_claim_outcome = if watcher_claimed {
-        WatcherHandoffClaimOutcome::Spawned
-    } else {
-        WatcherHandoffClaimOutcome::ReusedExisting
-    };
-    let _ = watcher_claim_replaced_existing;
-    if watcher_claimed {
-        #[cfg(unix)]
-        {
-            let on_standby = shared_owned.http.cached_serenity_ctx.get().is_none();
-            if on_standby {
-                let ts = chrono::Local::now().format("%H:%M:%S");
-                tracing::info!(
-                    "  [{ts}] ⏭ standby relay: skipping tmux watcher spawn for channel {}; spawning JSONL→Discord standby_relay",
-                    channel_id
-                );
-                let _ = shared_owned.tmux_watchers.remove(watcher_owner_channel_id);
-                if let Some(http_for_standby) = shared_owned.serenity_http_or_token_fallback() {
-                    let placeholder_msg_id_opt = if inflight_state.current_msg_id == 0 {
-                        None
-                    } else {
-                        Some(serenity::MessageId::new(inflight_state.current_msg_id))
-                    };
-                    let output_path_for_standby = output_path.clone();
-                    let turn_binding_for_standby =
-                        super::standby_relay::StandbyRelayTurnBinding::from_state(&inflight_state);
-                    let cancel_for_standby = Arc::new(std::sync::atomic::AtomicBool::new(false));
-                    let shared_for_standby = shared_owned.clone();
-                    let provider_for_standby = provider.clone();
-                    super::task_supervisor::spawn_observed(
-                        "turn_bridge_standby_relay",
-                        super::standby_relay::run_standby_relay(
-                            http_for_standby,
-                            channel_id,
-                            placeholder_msg_id_opt,
-                            output_path_for_standby,
-                            turn_binding_for_standby,
-                            last_offset,
-                            cancel_for_standby,
-                            shared_for_standby,
-                            provider_for_standby,
-                            // #2448: bumped from 900s (15min) heuristic stop
-                            // signal to a 1800s (30min) safety backstop. The
-                            // authoritative exit signal is now
-                            // `InflightSignal::Completed`, broadcast by
-                            // `CompletionGuard` on bridge drop.
-                            std::time::Duration::from_secs(1800),
-                        ),
-                    );
-                    *standby_relay_owns_output = true;
-                    inflight_state
-                        .set_relay_owner_kind(super::inflight::RelayOwnerKind::StandbyRelay);
-                    // #2263: intentionally leave `watcher_owns_live_relay = false`
-                    // on the standby branch.
-                    //
-                    // The flag's downstream contract in
-                    // `tmux::watcher_should_yield_to_inflight_state` is
-                    // narrowly "the restored TMUX WATCHER itself owns
-                    // delivery for this turn — do not yield". The standby
-                    // branch never spawns a watcher (the briefly-claimed
-                    // slot was just removed at line ~1477); the
-                    // `standby_relay` task is a separate, non-persisted
-                    // delivery owner whose ownership is NOT representable
-                    // by this single boolean.
-                    //
-                    // Setting the flag to `true` here would over-claim
-                    // ownership for any watcher restored against this
-                    // state on a different node (or after failover) — it
-                    // would short-circuit the yield gate and let a
-                    // restored watcher deliver concurrently with a still-
-                    // alive standby_relay, producing duplicate Discord
-                    // posts (codex adversarial review on #2263).
-                    //
-                    // The cost of keeping it `false` is the phantom-
-                    // bridge yield window: on restart, a restored watcher
-                    // whose tmux offset overlaps `turn_start_offset` will
-                    // yield to a bridge owner that died with the original
-                    // standby process and will suppress relay for the
-                    // overlapping batch. The inflight row is then cleared
-                    // by the `INFLIGHT_STALENESS_THRESHOLD_SECS` (300s)
-                    // staleness path in `classify_inflight_diagnostic_state`
-                    // (router/message_handler.rs) and the recovery-engine
-                    // sweep, after which a follow-up user turn proceeds
-                    // normally. The completed standby_relay response that
-                    // landed before the crash is preserved on Discord (it
-                    // was posted before the process died); the failure
-                    // mode is the user-visible stall on the FOLLOW-UP
-                    // turn until staleness sweep, NOT a dropped response.
-                    //
-                    // #2376 records `relay_owner_kind = standby_relay` so a
-                    // restored watcher can yield for every live batch, not
-                    // only batches that overlap the original turn_start_offset.
-                    // A future owner-lease timestamp can distinguish
-                    // dead-standby from live-standby and remove the phantom
-                    // yield window entirely.
-                    //
-                    // Per-turn in-process state is still correctly tracked
-                    // by `standby_relay_owns_output = true` above; that
-                    // local flag is what gates the bridge's terminal
-                    // delivery suppression for the current turn.
-                    let _ = save_inflight_state(inflight_state);
-                } else {
-                    let ts = chrono::Local::now().format("%H:%M:%S");
-                    tracing::warn!(
-                        "  [{ts}] ⚠ standby relay skipped: no Http source for channel {}",
-                        channel_id
-                    );
-                }
-            } else if let Some(http_bg) = shared_owned.serenity_http_or_token_fallback() {
-                let shared_bg = shared_owned.clone();
-                inflight_state.set_relay_owner_kind(super::inflight::RelayOwnerKind::Watcher);
-                let restored_turn = super::tmux::restored_watcher_turn_from_inflight(
-                    inflight_state,
-                    &tmux_session_name,
-                    true,
-                );
-                if let Ok(mut guard) = resume_offset.lock() {
-                    *guard = Some(last_offset);
-                }
-                turn_delivered.store(false, std::sync::atomic::Ordering::Relaxed);
-                if watcher_claim_replaced_existing {
-                    shared_owned.record_tmux_watcher_reconnect(channel_id);
-                }
-                super::task_supervisor::spawn_observed_tmux_watcher(
-                    "turn_bridge_tmux_output_watcher_with_restore",
-                    shared_bg.clone(),
-                    tmux_session_name.clone(),
-                    cancel.clone(),
-                    super::tmux::tmux_output_watcher_with_restore(
-                        channel_id,
-                        http_bg,
-                        shared_bg,
-                        output_path,
-                        tmux_session_name,
-                        last_offset,
-                        cancel,
-                        paused,
-                        resume_offset,
-                        pause_epoch,
-                        turn_delivered,
-                        last_heartbeat_ts_ms,
-                        restored_turn,
-                    ),
-                );
-                *watcher_relay_available_for_turn = true;
-                let _ = save_inflight_state(inflight_state);
-                watcher_ready_for_relay = true;
-            } else {
-                let ts = chrono::Local::now().format("%H:%M:%S");
-                tracing::warn!(
-                    "  [{ts}] ⚠ no Http source (neither cached_serenity_ctx nor cached_bot_token); tmux watcher not started for channel {}",
-                    channel_id
-                );
-                if let Some((_, handle)) =
-                    shared_owned.tmux_watchers.remove(watcher_owner_channel_id)
-                {
-                    handle
-                        .cancel
-                        .store(true, std::sync::atomic::Ordering::Relaxed);
-                }
-            }
-        }
-    }
-    if watcher_ready_for_relay {
-        *tmux_handed_off = true;
-        inflight_state.set_relay_owner_kind(super::inflight::RelayOwnerKind::Watcher);
-        *watcher_owns_assistant_relay = true;
-        let _ = save_inflight_state(inflight_state);
-        if let Some(watcher) = shared_owned.tmux_watchers.get(watcher_owner_channel_id) {
-            *watcher_relay_available_for_turn = true;
-            if let Ok(mut guard) = watcher.resume_offset.lock() {
-                *guard = Some(last_offset);
-            }
-            watcher
-                .turn_delivered
-                .store(false, std::sync::atomic::Ordering::Relaxed);
-            // #3016 phase 2: register the turn with the single-authority
-            // finalizer BEFORE unpausing the watcher. Message arrival order in
-            // the actor replaces the deleted Release/AcqRel ordering: the
-            // ledger now knows the turn exists (with the watcher as relay
-            // owner) before the watcher can submit its terminal. The ledger is
-            // the authority that superseded the legacy `mailbox_finalize_owed`
-            // flag (removed in #3016 phase-5b2) and the CAS revoke deleted from
-            // the bridge finalize branches below.
-            shared_owned.turn_finalizer.register_start(
-                super::turn_finalizer::TurnKey::new(
-                    channel_id,
-                    inflight_state.effective_finalizer_turn_id(),
-                    shared_owned.restart.current_generation,
-                ),
-                provider.clone(),
-                super::inflight::RelayOwnerKind::Watcher,
-                // #3016 phase-5a: prime the reconcile cache at register time.
-                shared_owned,
-            );
-            watcher
-                .paused
-                .store(false, std::sync::atomic::Ordering::Release);
-        }
-    }
-    *state_dirty = true;
-    if done {
-        *terminal_control_drain_until = None;
-    }
 }
 
 pub(super) fn spawn_turn_bridge(
@@ -2775,471 +2475,68 @@ pub(super) fn spawn_turn_bridge(
                             tmux_session_name,
                             last_offset,
                         } => {
-                            terminal_control_ready_observed = true;
-                            tmux_last_offset = Some(last_offset);
-                            inflight_state.runtime_kind =
-                                Some(RuntimeHandoffKind::LegacyTmuxWrapper);
-                            inflight_state.tmux_session_name = Some(tmux_session_name.clone());
-                            inflight_state.output_path = Some(output_path.clone());
-                            inflight_state.input_fifo_path =
-                                Some(input_fifo_path).filter(|path| !path.is_empty());
-                            inflight_state.last_offset = last_offset;
-
-                            // #226: Atomic claim via try_claim_watcher
-                            let cancel = Arc::new(std::sync::atomic::AtomicBool::new(false));
-                            let paused = Arc::new(std::sync::atomic::AtomicBool::new(true));
-                            let resume_offset = Arc::new(std::sync::Mutex::new(None::<u64>));
-                            let pause_epoch = Arc::new(std::sync::atomic::AtomicU64::new(1));
-                            let turn_delivered =
-                                Arc::new(std::sync::atomic::AtomicBool::new(false));
-                            let last_heartbeat_ts_ms =
-                                Arc::new(std::sync::atomic::AtomicI64::new(
-                                    super::tmux_watcher_now_ms(),
-                                ));
-                            let handle = TmuxWatcherHandle {
-                                tmux_session_name: tmux_session_name.clone(),
-                                output_path: output_path.clone(),
-                                paused: paused.clone(),
-                                resume_offset: resume_offset.clone(),
-                                cancel: cancel.clone(),
-                                pause_epoch: pause_epoch.clone(),
-                                turn_delivered: turn_delivered.clone(),
-                                last_heartbeat_ts_ms: last_heartbeat_ts_ms.clone(),
-                            };
-                            #[cfg(unix)]
-                            let (watcher_claimed, watcher_claim_replaced_existing) = {
-                                // #1135: Reuse a live watcher for the same
-                                // tmux session; replace only stale or
-                                // different-session incumbents.
-                                let claim = super::tmux::claim_or_reuse_watcher(
-                                    &shared_owned.tmux_watchers,
-                                    channel_id,
-                                    handle,
-                                    &provider,
-                                    "turn_bridge_tmux_ready",
-                                );
-                                watcher_owner_channel_id = claim.owner_channel_id();
-                                let _ = inflight_state.set_watcher_owner_channel_id(watcher_owner_channel_id.get());
-                                (claim.should_spawn(), claim.replaced_existing())
-                            };
-                            #[cfg(not(unix))]
-                            let (watcher_claimed, watcher_claim_replaced_existing) = {
-                                let _ = handle;
-                                (false, false)
-                            };
-                            #[cfg(unix)]
-                            let mut watcher_ready_for_relay = !watcher_claimed;
-                            #[cfg(not(unix))]
-                            let mut watcher_ready_for_relay = false;
-                            watcher_handoff_claim_outcome = if watcher_claimed {
-                                WatcherHandoffClaimOutcome::Spawned
-                            } else {
-                                WatcherHandoffClaimOutcome::ReusedExisting
-                            };
-                            if watcher_claimed {
-                                #[cfg(unix)]
-                                {
-                                    // Phase 5.3 of intake-node-routing
-                                    // (issue #2011): on cluster-standby nodes
-                                    // (no Discord gateway lease, no
-                                    // `cached_serenity_ctx`), bypass the tmux
-                                    // watcher entirely — its internal state
-                                    // machine has multiple gateway-coupled
-                                    // assumptions that prevent the relay step
-                                    // from firing on standby (verified
-                                    // 2026-05-10). Instead, leave
-                                    // `watcher_relay_available_for_turn=false`
-                                    // so the bridge delivers the response
-                                    // itself via
-                                    // `gateway.replace_message_with_outcome`
-                                    // after the producer's `Done` event
-                                    // populates `delivery_response`. The
-                                    // bridge's REST gateway path already uses
-                                    // `serenity_http_or_token_fallback()`
-                                    // (Phase 5.2) so the post lands on Discord
-                                    // even without the gateway runtime.
-                                    //
-                                    // Leader path is unchanged: when
-                                    // `cached_serenity_ctx` is set, spawn the
-                                    // watcher as before so streaming partial
-                                    // output continues to work.
-                                    let on_standby = shared_owned.http.cached_serenity_ctx.get().is_none();
-                                    if on_standby {
-                                        // Phase 5.3 of intake-node-routing (issue #2011):
-                                        // skip the watcher entirely on standby and
-                                        // spawn the standalone JSONL → Discord relay
-                                        // task instead. The watcher's leader-only
-                                        // state machine prevents its relay step from
-                                        // firing on standby nodes; bypassing it
-                                        // sidesteps an entire class of
-                                        // gateway-coupling bugs.
-                                        let ts = chrono::Local::now().format("%H:%M:%S");
-                                        tracing::info!(
-                                            "  [{ts}] ⏭ standby relay: skipping tmux watcher spawn for channel {}; spawning JSONL→Discord standby_relay",
-                                            channel_id
-                                        );
-                                        // Drop the registered watcher slot so a
-                                        // subsequent turn does not falsely reuse
-                                        // a "live" watcher that we never spawned.
-                                        // Do NOT call `cancel.store(true)` on the
-                                        // returned handle: the inner cancel Arc
-                                        // is shared with the local `cancel` and
-                                        // would pre-cancel the standby_relay we
-                                        // are about to spawn (Codex P1 review on
-                                        // PR #2012). The cancel Arc is otherwise
-                                        // unused on this branch since no watcher
-                                        // task ever reads it.
-                                        let _ = shared_owned
-                                            .tmux_watchers
-                                            .remove(&watcher_owner_channel_id);
-                                        if let Some(http_for_standby) =
-                                            shared_owned.serenity_http_or_token_fallback()
-                                        {
-                                            let placeholder_msg_id_opt =
-                                                if inflight_state.current_msg_id == 0 {
-                                                    None
-                                                } else {
-                                                    Some(serenity::MessageId::new(
-                                                        inflight_state.current_msg_id,
-                                                    ))
-                                            };
-                                            let output_path_for_standby = output_path.clone();
-                                            let turn_binding_for_standby = super::standby_relay::StandbyRelayTurnBinding::from_state(
-                                                &inflight_state,
-                                            );
-                                            // Use a fresh cancel Arc, independent
-                                            // from the watcher's `cancel` (which
-                                            // is shared via `handle.cancel`).
-                                            let cancel_for_standby = Arc::new(
-                                                std::sync::atomic::AtomicBool::new(false),
-                                            );
-                                            let shared_for_standby = shared_owned.clone();
-                                            let provider_for_standby = provider.clone();
-                                            super::task_supervisor::spawn_observed(
-                                                "turn_bridge_runtime_standby_relay",
-                                                super::standby_relay::run_standby_relay(
-                                                http_for_standby,
-                                                channel_id,
-                                                placeholder_msg_id_opt,
-                                                output_path_for_standby,
-                                                turn_binding_for_standby,
-                                                last_offset,
-                                                cancel_for_standby,
-                                                shared_for_standby,
-                                                provider_for_standby,
-                                                // #2448: see TmuxReady branch
-                                                // — timeout demoted to safety
-                                                // backstop after broadcast
-                                                // exit signal landed.
-                                                std::time::Duration::from_secs(1800),
-                                                ),
-                                            );
-                                            standby_relay_owns_output = true;
-                                            inflight_state.set_relay_owner_kind(
-                                                super::inflight::RelayOwnerKind::StandbyRelay,
-                                            );
-                                            // #2263: see the helper-fn
-                                            // `handle_watcher_runtime_handoff`
-                                            // standby branch — intentionally
-                                            // leave `watcher_owns_live_relay = false`
-                                            // because the standby_relay task
-                                            // is not a tmux watcher, and the
-                                            // yield-gate flag would over-claim
-                                            // ownership for a watcher restored
-                                            // by a different node, risking
-                                            // duplicate Discord delivery.
-                                            // Per-turn delivery ownership is
-                                            // tracked both locally by
-                                            // `standby_relay_owns_output` and
-                                            // durably by `relay_owner_kind`.
-                                            let _ = save_inflight_state(&inflight_state);
-                                        } else {
-                                            let ts = chrono::Local::now().format("%H:%M:%S");
-                                            tracing::warn!(
-                                                "  [{ts}] ⚠ standby relay skipped: no Http source for channel {}",
-                                                channel_id
-                                            );
-                                        }
-                                        // Leave watcher_relay_available_for_turn=false
-                                        // and watcher_ready_for_relay=false so the
-                                        // bridge does NOT delegate to a non-existent
-                                        // watcher. The standby_relay task delivers
-                                        // the response independently.
-                                    } else if let Some(http_bg) = shared_owned.serenity_http_or_token_fallback() {
-                                        let shared_bg = shared_owned.clone();
-                                        inflight_state.set_relay_owner_kind(
-                                            super::inflight::RelayOwnerKind::Watcher,
-                                        );
-                                        let restored_turn =
-                                            super::tmux::restored_watcher_turn_from_inflight(
-                                                &inflight_state,
-                                                &tmux_session_name,
-                                                true,
-                                            );
-                                        if let Ok(mut guard) = resume_offset.lock() {
-                                            *guard = Some(last_offset);
-                                        }
-                                        turn_delivered.store(false, Ordering::Relaxed);
-                                        if watcher_claim_replaced_existing {
-                                            shared_owned.record_tmux_watcher_reconnect(channel_id);
-                                        }
-                                        super::task_supervisor::spawn_observed_tmux_watcher(
-                                            "turn_bridge_runtime_tmux_output_watcher_with_restore",
-                                            shared_bg.clone(),
-                                            tmux_session_name.clone(),
-                                            cancel.clone(),
-                                            super::tmux::tmux_output_watcher_with_restore(
-                                            channel_id,
-                                            http_bg,
-                                            shared_bg,
-                                            output_path,
-                                            tmux_session_name,
-                                            last_offset,
-                                            cancel,
-                                            paused,
-                                            resume_offset,
-                                            pause_epoch,
-                                            turn_delivered,
-                                            last_heartbeat_ts_ms,
-                                            restored_turn,
-                                            ),
-                                        );
-                                        watcher_relay_available_for_turn = true;
-                                        let _ = save_inflight_state(&inflight_state);
-                                        watcher_ready_for_relay = true;
-                                    } else {
-                                        let ts = chrono::Local::now().format("%H:%M:%S");
-                                        tracing::warn!(
-                                            "  [{ts}] ⚠ no Http source (neither cached_serenity_ctx nor cached_bot_token); tmux watcher not started for channel {}",
-                                            channel_id
-                                        );
-                                        if let Some((_, handle)) =
-                                            shared_owned
-                                                .tmux_watchers
-                                                .remove(&watcher_owner_channel_id)
-                                        {
-                                            handle.cancel.store(true, Ordering::Relaxed);
-                                        }
-                                    }
-                                }
-                            }
-                            if watcher_ready_for_relay {
-                                tmux_handed_off = true;
-                                inflight_state.set_relay_owner_kind(
-                                    super::inflight::RelayOwnerKind::Watcher,
-                                );
-                                watcher_owns_assistant_relay = true;
-                                let _ = save_inflight_state(&inflight_state);
-                                if let Some(watcher) =
-                                    shared_owned.tmux_watchers.get(&watcher_owner_channel_id)
-                                {
-                                    watcher_relay_available_for_turn = true;
-                                    if let Ok(mut guard) = watcher.resume_offset.lock() {
-                                        *guard = Some(last_offset);
-                                    }
-                                    watcher.turn_delivered.store(false, Ordering::Relaxed);
-                                    // #1452 (Codex P1): publish the mailbox-finalization
-                                    // debt BEFORE unpausing the watcher.
-                                    //
-                                    // The watcher's terminal `swap(false, AcqRel)` runs
-                                    // as soon as it sees a Done event; if we delayed
-                                    // the store until the bridge's later delegation
-                                    // decision (line 2419+), the watcher could swap
-                                    // first, observe `false`, skip `mailbox_finish_turn`,
-                                    // and the bridge's late `store(true)` would leave
-                                    // stale debt that either keeps `cancel_token`
-                                    // permanently set OR is consumed by a future
-                                    // watcher event for the WRONG turn.
-                                    //
-                                    // #3016 phase-5b2: the legacy
-                                    // `mailbox_finalize_owed` store that used to
-                                    // publish "bridge will delegate finalization"
-                                    // here is removed; the `register_start` below
-                                    // (RelayOwnerKind::Watcher) is the ledger
-                                    // authority that replaced it.
-                                    // #3016 phase 3: register the turn with the
-                                    // single-authority finalizer BEFORE
-                                    // unpausing the watcher — same as the
-                                    // `handle_watcher_runtime_handoff` helper.
-                                    // This legacy `StreamMessage::TmuxReady`
-                                    // handoff does NOT go through that helper, so
-                                    // without this the watcher terminal would
-                                    // have no Watcher-owned ledger entry — and
-                                    // a busy-pane gate-timeout would finalize
-                                    // immediately instead of arming the
-                                    // deadline-backstop. Registering here with
-                                    // the same finalizer id makes it defer.
-                                    shared_owned.turn_finalizer.register_start(
-                                        super::turn_finalizer::TurnKey::new(
-                                            channel_id,
-                                            inflight_state.effective_finalizer_turn_id(),
-                                            shared_owned.restart.current_generation,
-                                        ),
-                                        provider.clone(),
-                                        super::inflight::RelayOwnerKind::Watcher,
-                                        // #3016 phase-5a: prime the reconcile cache
-                                        // at register time.
-                                        &shared_owned,
-                                    );
-                                    // #1452 (Codex iter 3 P1) / #3016 phase-5b2:
-                                    // unpause uses Release ordering so a watcher
-                                    // observing `paused = false` is guaranteed to
-                                    // also observe the prior writes — the
-                                    // `register_start` (RelayOwnerKind::Watcher)
-                                    // ledger entry that now drives the
-                                    // gate-timeout defer. With Relaxed ordering on
-                                    // a weakly-ordered platform the writes could
-                                    // be reordered, letting the watcher unpause
-                                    // and submit a terminal before the ledger
-                                    // knows the turn exists.
-                                    watcher.paused.store(false, Ordering::Release);
-                                }
-                            }
-                            state_dirty = true;
-                            if done {
-                                terminal_control_drain_until = None;
-                            }
-                        }
-                        StreamMessage::RuntimeReady { handoff } => {
-                            terminal_control_ready_observed = true;
-                            match handoff {
-                                RuntimeHandoff::LegacyTmuxWrapper {
+                            let outcome = handle_runtime_handoff_loop_message(
+                                RuntimeHandoffLoopMessage::TmuxReady {
                                     output_path,
                                     input_fifo_path,
                                     tmux_session_name,
                                     last_offset,
-                                } => {
-                                handle_watcher_runtime_handoff(
-                                    &shared_owned,
-                                    &provider,
+                                },
+                                RuntimeHandoffLoopContext {
+                                    shared_owned: &shared_owned,
+                                    provider: &provider,
                                     channel_id,
-                                    &mut inflight_state,
-                                    RuntimeHandoffKind::LegacyTmuxWrapper,
-                                    output_path,
-                                    Some(input_fifo_path),
-                                    tmux_session_name,
-                                    last_offset,
-                                    &mut tmux_last_offset,
-                                    &mut watcher_owner_channel_id,
-                                    &mut standby_relay_owns_output,
-                                    &mut watcher_relay_available_for_turn,
-                                    &mut watcher_handoff_claim_outcome,
-                                    &mut tmux_handed_off,
-                                    &mut watcher_owns_assistant_relay,
-                                    &mut state_dirty,
                                     done,
-                                    &mut terminal_control_drain_until,
-                                );
+                                    adk_session_name: &adk_session_name,
+                                },
+                                RuntimeHandoffLoopState {
+                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
+                                    tmux_last_offset: &mut tmux_last_offset,
+                                    inflight_state: &mut inflight_state,
+                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
+                                    standby_relay_owns_output: &mut standby_relay_owns_output,
+                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
+                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
+                                    tmux_handed_off: &mut tmux_handed_off,
+                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
+                                    state_dirty: &mut state_dirty,
+                                    terminal_control_drain_until: &mut terminal_control_drain_until,
+                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
+                                },
+                            )
+                            .await;
+                            match outcome {
+                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
                             }
-                            RuntimeHandoff::ClaudeTui {
-                                transcript_path,
-                                tmux_session_name,
-                                last_offset,
-                            } => {
-                                handle_watcher_runtime_handoff(
-                                    &shared_owned,
-                                    &provider,
+                        }
+                        StreamMessage::RuntimeReady { handoff } => {
+                            let outcome = handle_runtime_handoff_loop_message(
+                                RuntimeHandoffLoopMessage::RuntimeReady { handoff },
+                                RuntimeHandoffLoopContext {
+                                    shared_owned: &shared_owned,
+                                    provider: &provider,
                                     channel_id,
-                                    &mut inflight_state,
-                                    RuntimeHandoffKind::ClaudeTui,
-                                    transcript_path,
-                                    None,
-                                    tmux_session_name,
-                                    last_offset,
-                                    &mut tmux_last_offset,
-                                    &mut watcher_owner_channel_id,
-                                    &mut standby_relay_owns_output,
-                                    &mut watcher_relay_available_for_turn,
-                                    &mut watcher_handoff_claim_outcome,
-                                    &mut tmux_handed_off,
-                                    &mut watcher_owns_assistant_relay,
-                                    &mut state_dirty,
                                     done,
-                                    &mut terminal_control_drain_until,
-                                );
-                            }
-                            RuntimeHandoff::CodexTui {
-                                rollout_path,
-                                thread_id,
-                                tmux_session_name,
-                                last_offset,
-                            } => {
-                                if let Some(thread_id) = thread_id {
-                                    inflight_state.session_id = Some(thread_id);
-                                }
-                                handle_watcher_runtime_handoff(
-                                    &shared_owned,
-                                    &provider,
-                                    channel_id,
-                                    &mut inflight_state,
-                                    RuntimeHandoffKind::CodexTui,
-                                    rollout_path,
-                                    None,
-                                    tmux_session_name,
-                                    last_offset,
-                                    &mut tmux_last_offset,
-                                    &mut watcher_owner_channel_id,
-                                    &mut standby_relay_owns_output,
-                                    &mut watcher_relay_available_for_turn,
-                                    &mut watcher_handoff_claim_outcome,
-                                    &mut tmux_handed_off,
-                                    &mut watcher_owns_assistant_relay,
-                                    &mut state_dirty,
-                                    done,
-                                    &mut terminal_control_drain_until,
-                                );
-                            }
-                            RuntimeHandoff::ProcessBackend {
-                                output_path,
-                                session_name,
-                                last_offset,
-                            } => {
-                                tmux_last_offset = Some(last_offset);
-                                inflight_state.runtime_kind =
-                                    Some(RuntimeHandoffKind::ProcessBackend);
-                                inflight_state.tmux_session_name = Some(session_name);
-                                inflight_state.output_path = Some(output_path);
-                                inflight_state.input_fifo_path = None;
-                                inflight_state.last_offset = last_offset;
-                                state_dirty = true;
-                                // #2235: see CodexTui arm — durable stamp of
-                                // runtime_kind across a bridge-crash window.
-                                let _ = save_inflight_state(&inflight_state);
-                                if done {
-                                    terminal_control_drain_until = None;
-                                }
-                                }
-                            RuntimeHandoff::ClaudeEAdapter {
-                                output_path,
-                                session_name,
-                                last_offset,
-                            } => {
-                                // Phase 1 of the claude-e rollout (see
-                                // `docs/claude-e-rollout/`). The adapter
-                                // is a per-turn PTY spawn — no tmux pane
-                                // backs it, so `tmux_session_name` must
-                                // stay `None` to satisfy the
-                                // `inflight_tmux_one_to_one` invariant
-                                // when a channel switches between TUI
-                                // and claude-e. `session_name` is the
-                                // logical adapter id (Claude session uuid
-                                // or `claude-e-{pid}`); it does not map
-                                // to a tmux pane and is intentionally
-                                // not stamped here.
-                                let _ = session_name;
-                                tmux_last_offset = Some(last_offset);
-                                inflight_state.runtime_kind =
-                                    Some(RuntimeHandoffKind::ClaudeEAdapter);
-                                inflight_state.tmux_session_name = None;
-                                inflight_state.output_path = Some(output_path);
-                                inflight_state.input_fifo_path = None;
-                                inflight_state.last_offset = last_offset;
-                                state_dirty = true;
-                                let _ = save_inflight_state(&inflight_state);
-                                if done {
-                                    terminal_control_drain_until = None;
-                                }
-                            }
+                                    adk_session_name: &adk_session_name,
+                                },
+                                RuntimeHandoffLoopState {
+                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
+                                    tmux_last_offset: &mut tmux_last_offset,
+                                    inflight_state: &mut inflight_state,
+                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
+                                    standby_relay_owns_output: &mut standby_relay_owns_output,
+                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
+                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
+                                    tmux_handed_off: &mut tmux_handed_off,
+                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
+                                    state_dirty: &mut state_dirty,
+                                    terminal_control_drain_until: &mut terminal_control_drain_until,
+                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
+                                },
+                            )
+                            .await;
+                            match outcome {
+                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
                             }
                         }
                         StreamMessage::ProcessReady {
@@ -3247,40 +2544,68 @@ pub(super) fn spawn_turn_bridge(
                             session_name,
                             last_offset,
                         } => {
-                            terminal_control_ready_observed = true;
-                            // ProcessBackend completed first turn.
-                            // No tmux watcher needed — process sessions are monitored
-                            // inline via SessionProbe::process during read_output_file_until_result.
-                            // Do NOT set tmux_handed_off: ProcessBackend has no watcher,
-                            // so the handoff cleanup path would delete the placeholder
-                            // with no one to send the final response.
-                            tmux_last_offset = Some(last_offset);
-                            inflight_state.runtime_kind = Some(RuntimeHandoffKind::ProcessBackend);
-                            inflight_state.tmux_session_name = Some(session_name);
-                            inflight_state.output_path = Some(output_path);
-                            inflight_state.input_fifo_path = None;
-                            inflight_state.last_offset = last_offset;
-                            state_dirty = true;
-                            // #2235: persist runtime_kind stamp immediately —
-                            // ProcessBackend has no watcher so we want the
-                            // on-disk row to reflect the new backend before
-                            // any potential bridge crash.
-                            let _ = save_inflight_state(&inflight_state);
-                            if done {
-                                terminal_control_drain_until = None;
+                            let outcome = handle_runtime_handoff_loop_message(
+                                RuntimeHandoffLoopMessage::ProcessReady {
+                                    output_path,
+                                    session_name,
+                                    last_offset,
+                                },
+                                RuntimeHandoffLoopContext {
+                                    shared_owned: &shared_owned,
+                                    provider: &provider,
+                                    channel_id,
+                                    done,
+                                    adk_session_name: &adk_session_name,
+                                },
+                                RuntimeHandoffLoopState {
+                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
+                                    tmux_last_offset: &mut tmux_last_offset,
+                                    inflight_state: &mut inflight_state,
+                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
+                                    standby_relay_owns_output: &mut standby_relay_owns_output,
+                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
+                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
+                                    tmux_handed_off: &mut tmux_handed_off,
+                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
+                                    state_dirty: &mut state_dirty,
+                                    terminal_control_drain_until: &mut terminal_control_drain_until,
+                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
+                                },
+                            )
+                            .await;
+                            match outcome {
+                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
                             }
                         }
                         StreamMessage::OutputOffset { offset } => {
-                            tmux_last_offset = Some(offset);
-                            inflight_state.last_offset = offset;
-                            maybe_refresh_active_turn_activity_heartbeat(
-                                shared_owned.as_ref(),
-                                &provider,
-                                &inflight_state,
-                                adk_session_name.as_deref(),
-                                &mut last_activity_heartbeat_at,
-                            );
-                            state_dirty = true;
+                            let outcome = handle_runtime_handoff_loop_message(
+                                RuntimeHandoffLoopMessage::OutputOffset { offset },
+                                RuntimeHandoffLoopContext {
+                                    shared_owned: &shared_owned,
+                                    provider: &provider,
+                                    channel_id,
+                                    done,
+                                    adk_session_name: &adk_session_name,
+                                },
+                                RuntimeHandoffLoopState {
+                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
+                                    tmux_last_offset: &mut tmux_last_offset,
+                                    inflight_state: &mut inflight_state,
+                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
+                                    standby_relay_owns_output: &mut standby_relay_owns_output,
+                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
+                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
+                                    tmux_handed_off: &mut tmux_handed_off,
+                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
+                                    state_dirty: &mut state_dirty,
+                                    terminal_control_drain_until: &mut terminal_control_drain_until,
+                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
+                                },
+                            )
+                            .await;
+                            match outcome {
+                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
+                            }
                         }
                         }
                     }
